scripts/requirements.txt
transformers==4.25.1
huggingface-hub==0.10.0
pyyaml==5.4
regex!=2019.12.17
tokenizers==0.13.0
scikit-learn==1.2.0
joblib==1.1.1
pinecone-client==2.2.1
matplotlib==3.7.1
contourpy==1.0.1
cycler==0.10
kiwisolver==1.0.1
Pillow==9.4.0
pyparsing==2.3.1
argilla==1.4.0
numpy==1.23.0
Flask==2.0.1
Flask-Cors==3.0.10
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-auth-httplib2==0.1.0
google-api-python-client==2.18.0
requests==2.28.1
requests-oauthlib==1.3.1
cryptography==3.4.8
airtable-python-wrapper==0.15.1
python-dotenv==0.19.1
cloudinary==1.26.0
openai==0.27.2
urllib3==1.26.13
charset-normalizer==2.1.1
chardet==5.1.0
bs4==0.0.1
